requirements.txt
asyncio
fugle_marketdata
confluent_kafka==1.9.2
requests
aiohttp
# asyncpg
pyspark==3.5.0
aiokafka